Add unifi controller setup for provider test
This commit builds on top of the test setup of terraform-provider-unifi.
Running the tests of the terraform provider, I experienced issues like
the controller not being properly setup with a user. By using compose V2
to start the unifi controller and python and urllib for user setup, I
hope to mitigate these issues.